add RSSI return and fix SSID and BSSID methods
fixes: https://github.com/espressif/arduino-esp32/issues/144

+++ libraries/WiFi/src/WiFiSTA.h
@@ -78,7 +78,7 @@ public:
     uint8_t * BSSID();
     String BSSIDstr();
 
-    int32_t RSSI();
+    int8_t RSSI();
 
     static void _setStatus(wl_status_t status);
 protected:
